A list of over 100k valid, unique, recognizable identifiers
Every line of text in this file is a unique sequence of 2 to 15
lowercase ASCII characters. Most will be recognized by English
speakers as given names.

This is simply allnames.txt (every name, sorted by which name was
given to the greatest number of babies born in a year), except with
two changes:

1) The "NATO Phonetic Alphabet" has prepended at the beginning
2) Names are lower case

While this file could be handy for many things, the idea is that a
project which is trying to deobfuscate code can simply grab an
identifier from this list to use as functions names. This will
hopefully make reading the source code easier.

Note that no attempt has been made to remove homophones (names that
are spelled differently but sound similar).